Moved Config to a class Additional tests of authentication, resources

+++ resources/store.py
@@ -1,4 +1,6 @@
 from flask_restful import Resource
+from flask_jwt_extended import jwt_required, get_jwt
+
 from models.store import StoreModel
 
 
@@ -9,6 +11,7 @@ class Store(Resource):
             return store.json()
         return {"message": "Store not found."}, 404
 
+    @jwt_required()
     def post(self, name):
         if StoreModel.find_by_name(name):
             return {"message": f"A store with name '{name}' already exists."}
@@ -20,7 +23,12 @@ class Store(Resource):
 
         return store.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        # Using claims
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         store =  StoreModel.find_by_name(name)
         if store:
             store.delete_from_db()
